requirements.txt
openai
python-dotenv
langgraph
duckduckgo-search
langchain
ddgs
langchain_openai
fastapi
uvicorn